Enable aggregation of zarr datasets.
This feature enables the ability to let users "order" zarr stores that
contain aggregated datasets. Occording to an aggregation plan users can
combine the input data that into one zarr store.

+++ freva-client/src/freva_client/utils/types.py
@@ -9,5 +9,5 @@ class ZarrOptionsDict(TypedDict):
     public: bool
     """Flag indicating whether or not we need to create a public zarr ulr."""
 
-    ttl_seconds: int
+    ttl_seconds: float
     """TTL of the link in seconds."""
